added Micro-XRCE client to setup

+++ scripts/ros2_setup_packages.bash
@@ -8,28 +8,53 @@ aptinstall ros-jazzy-image-transport-plugins
 aptinstall v4l-utils
 
 echo "INFO: installing libcamera binary for Pi 5"
-aptinstall ppa:marco-sonic/rasppios
+sudo add-apt-repository ppa:marco-sonic/rasppios
+sudo apt update && $ sudo apt upgrade
 aptinstall libcamera-tools 
-aptisntall rpicam-apps-lite 
+aptinstall rpicam-apps-lite 
 aptinstall python3-picamera2
 
 U=${USER}
 sudo adduser $U video
 
 
-sudo apt -y install python3-colcon-meson
+aptinstall python3-colcon-meson
 mdkir -p ${PPWD}/src
-cd ${PPWD}/src
-git clone https://github.com/christianrauch/camera_ros.git
+
+# referenced from https://github.com/ARLunan/Raspberry-Pi-Camera-ROS/blob/main/Documents/RPCamera-Installation-RP5.md
+echo "INFO: installing camera_ros"
+if [[ ! -d  ${PPWD}/src/camera_ros ]]; then
+    cd ${PPWD}/src
+    git clone https://github.com/christianrauch/camera_ros.git
+else
+    cd ${PPWD}/src/camera_ros
+    git pull
+fi
 
 cd ${PPWD}
 source /opt/ros/$ROS_DISTRO/setup.bash
-rosdep install -y --from-paths src --ignore-src --rosdistro $ROS_DISTRO --skip-keys=libcamera
-sudo apt remove ros-$ROS_DISTRO-ros-camera
+dpkg -l ros-$ROS_DISTRO-ros-camera
+if [[ $? -eq 0 ]]; then
+    sudo apt remove ros-$ROS_DISTRO-ros-camera
+fi
 cd ${PPWD}/src
-git_clone_pull 'https://github.com/christianrauch/camera_ros.git' camera_ros ${CAMERA_ROS_VERSION} main
+
+echo "INFO: setting up Micro-XRCE-DDS-Agent"
+# sudo apt install -y ros-$ROS_DISTRO-rmw-fastrtps-cpp
+source /opt/ros/ROS_DISTRO/setup.bash
+if [[ ! -d  ${PPWD}/src/Micro-XRCE-DDS-Agent ]]; then
+    cd ${PPWD}/src
+    git clone -b v2.4.3  https://github.com/eProsima/Micro-XRCE-DDS-Agent.git
+else
+    cd ${PPWD}/src/Micro-XRCE-DDS-Agent
+    git pull
+fi
+
+
 rosdep install -y --from-paths src --ignore-src --rosdistro $ROS_DISTRO --skip-keys=libcamera
 colcon build --event-handlers=console_direct+
+
+
 
 
 # CAMERA_ROS_VERSION='0.5.0'
@@ -65,8 +90,6 @@ colcon build --event-handlers=console_direct+
 # echo "INFO: refreshing camera_ros:${CAMERA_ROS_VERSION}"
 # git_clone_pull 'https://github.com/christianrauch/camera_ros.git' camera_ros ${CAMERA_ROS_VERSION} main
 
-# echo "INFO: refreshing Micro-XRCE-DDS-Agent:${XRCE_DSS_VERSION}"
-# git_clone_pull 'https://github.com/eProsima/Micro-XRCE-DDS-Agent.git' Micro-XRCE-DDS-Agent ${XRCE_DSS_VERSION} master
-# rosdep install -y --from-paths src --ignore-src --rosdistro $ROS_DISTRO --skip-keys=libcamera
+
 
 exit 0
